+++ benchmarks/kernels/bench_per_token_quant_fp8.py
@@ -7,7 +7,7 @@ import torch
 
 from vllm import _custom_ops as ops
 from vllm.config import CompilationConfig, VllmConfig, set_current_vllm_config
-from vllm.model_executor.layers.fp8_quantization import QuantFP8
+from vllm.model_executor.layers.quantization.input_quant_fp8 import QuantFP8
 from vllm.model_executor.layers.quantization.utils.quant_utils import GroupShape
 from vllm.triton_utils import triton
